+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.4
+
+- Update `README.md`
+- Fixed: calculator not exported
+- `Myanmar Thingyan` days retrieving.
+
 ## 0.0.3
 
 - Update `README.md`